requirements.txt
Flask>=3.0.0
line-bot-sdk>=3.8.0
python-dotenv>=1.0.0
APScheduler>=3.10.4
pytz>=2024.1
SQLAlchemy>=2.0.23
gunicorn>=21.2.0
psycopg2-binary>=2.9.10
requests>=2.31.0
cryptography>=42.0.0
